geo: group Hong Kong and Macau into China for the chn worldview
In the chn worldview, Hong Kong and Macau were rendered as their own
entities. Fold them into China so chn treats HK/Macau/Taiwan as part of
China (Taiwan is already merged in Natural Earth's chn source).

- Add a worldview absorption pass (geo_rasterizer/src/absorb.rs): it grafts
  an absorbed dependency's geometry onto its sovereign feature and drops the
  dependency. Applied inside parse_geojson so no consumer can forget it.
- Scope HKG/MAC registry refs to iso/usa; chn no longer lists them.
- Move geo_rasterizer inline unit tests (absorb/entities/registry/cache)
  into tests/.
