+++ tools/geo_rasterizer/src/bin/registry_gen.rs
@@ -52,7 +52,7 @@ fn default_countries(worldview: Worldview) -> PathBuf {
 
 /// Register every ADM0_A3 (and its continent) found in `path` under `worldview`.
 fn register_source(reg: &mut Registry, worldview: Worldview, path: &Path) -> Result<()> {
-    let features = parse_geojson(path)?;
+    let features = parse_geojson(path, worldview.spec().id)?;
     let mut items: Vec<(String, bool, geo_types::MultiPolygon<f64>)> = Vec::new();
     for f in &features {
         items.push((
