+++ tools/geo_rasterizer/src/entities.rs
@@ -140,68 +140,3 @@ pub fn assemble_entities(features: &[ParsedFeature], registry: &Registry) -> Res
         geometry_for_country,
     })
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::registry::{Entry, Registry};
-    use geo_types::{Coord, LineString, Polygon};
-
-    fn feat(adm0: &str, continent: &str) -> ParsedFeature {
-        let sq = Polygon::new(
-            LineString(vec![
-                Coord { x: 0.0, y: 0.0 },
-                Coord { x: 1.0, y: 0.0 },
-                Coord { x: 1.0, y: 1.0 },
-                Coord { x: 0.0, y: 0.0 },
-            ]),
-            vec![],
-        );
-        ParsedFeature {
-            adm0_a3: adm0.into(),
-            iso_a3: adm0.into(),
-            iso_a3_eh: adm0.into(),
-            name: adm0.into(),
-            feature_type: "Country".into(),
-            continent: continent.into(),
-            region_un: "Africa".into(),
-            geometry: MultiPolygon(vec![sq]),
-        }
-    }
-
-    fn reg() -> Registry {
-        Registry {
-            schema: 1,
-            continents: vec![Entry {
-                code: "AS".into(),
-                id: 5,
-                refs: Default::default(),
-            }],
-            countries: vec![Entry {
-                code: "AAA".into(),
-                id: 3,
-                refs: Default::default(),
-            }],
-        }
-    }
-
-    #[test]
-    fn ids_come_from_registry_not_position() {
-        let m = assemble_entities(&[feat("AAA", "Asia")], &reg()).unwrap();
-        let aaa = m
-            .entities
-            .iter()
-            .find(|e| e.canonical_code == "AAA")
-            .unwrap();
-        assert_eq!(aaa.id, GeoEntityId(3));
-        assert_eq!(aaa.parent_id, Some(GeoEntityId(5)));
-    }
-
-    #[test]
-    fn unknown_adm0_is_an_error() {
-        let err = assemble_entities(&[feat("ZZZ", "Asia")], &reg())
-            .unwrap_err()
-            .to_string();
-        assert!(err.contains("ZZZ"), "got: {err}");
-    }
-}
